functions/reviews-service/requirements.txt
flask
dnspython
pymongo
python-dotenv
